rename utils to fieldUtils
