+++ src/fieldAnalyser/fieldUtils.ts
